314 refactor transform folder (#320)
* [DLMED] refactor transforms into folders

* [DLMED] add IO folder
