+++ monai/transforms/__init__.py
@@ -10,5 +10,13 @@
 # limitations under the License.
 
 from .compose import *
-from .composables import *
-from .transforms import *
+from .croppad.array import *
+from .croppad.dictionary import *
+from .intensity.array import *
+from .intensity.dictionary import *
+from .io.array import *
+from .io.dictionary import *
+from .spatial.array import *
+from .spatial.dictionary import *
+from .utility.array import *
+from .utility.dictionary import *
