+++ monai/transforms/spatial/array.py
@@ -9,7 +9,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """
-A collection of "vanilla" transforms
+A collection of "vanilla" transforms for spatial operations
 https://github.com/Project-MONAI/MONAI/wiki/MONAI_Design
 """
 
@@ -17,17 +17,14 @@ import warnings
 import numpy as np
 import scipy.ndimage
 import nibabel as nib
-from PIL import Image
 import torch
-from torch.utils.data._utils.collate import np_str_obj_array_pattern
 from skimage.transform import resize
 
-from monai.data.utils import (get_random_patch, get_valid_patch_size, correct_nifti_header_if_necessary, zoom_affine,
-                              compute_shape_offset, to_affine_nd)
+from monai.data.utils import zoom_affine, compute_shape_offset, to_affine_nd
 from monai.networks.layers.simplelayers import GaussianFilter
 from monai.transforms.compose import Transform, Randomizable
-from monai.transforms.utils import (create_control_grid, create_grid, create_rotate, create_scale, create_shear,
-                                    create_translate, rescale_array, generate_spatial_bounding_box)
+from monai.transforms.utils import create_control_grid, create_grid, create_rotate, \
+    create_scale, create_shear, create_translate
 from monai.utils.misc import ensure_tuple
 
 
@@ -170,258 +167,6 @@ class Orientation(Transform):
         new_affine = affine_ @ nib.orientations.inv_ornt_aff(spatial_ornt, shape)
         new_affine = to_affine_nd(affine, new_affine)
         return data_array, affine, new_affine
-
-
-class LoadNifti(Transform):
-    """
-    Load Nifti format file or files from provided path. If loading a list of
-    files, stack them together and add a new dimension as first dimension, and
-    use the meta data of the first image to represent the stacked result. Note
-    that the affine transform of all the images should be same if ``image_only=False``.
-    """
-
-    def __init__(self, as_closest_canonical=False, image_only=False, dtype=np.float32):
-        """
-        Args:
-            as_closest_canonical (bool): if True, load the image as closest to canonical axis format.
-            image_only (bool): if True return only the image volume, otherwise return image data array and header dict.
-            dtype (np.dtype, optional): if not None convert the loaded image to this data type.
-
-        Note:
-            The transform returns image data array if `image_only` is True,
-            or a tuple of two elements containing the data array, and the Nifti
-            header in a dict format otherwise.
-            if a dictionary header is returned:
-
-            - header['affine'] stores the affine of the image.
-            - header['original_affine'] will be additionally created to store the original affine.
-        """
-        self.as_closest_canonical = as_closest_canonical
-        self.image_only = image_only
-        self.dtype = dtype
-
-    def __call__(self, filename):
-        """
-        Args:
-            filename (str, list, tuple, file): path file or file-like object or a list of files.
-        """
-        filename = ensure_tuple(filename)
-        img_array = list()
-        compatible_meta = dict()
-        for name in filename:
-            img = nib.load(name)
-            img = correct_nifti_header_if_necessary(img)
-            header = dict(img.header)
-            header['filename_or_obj'] = name
-            header['affine'] = img.affine
-            header['original_affine'] = img.affine.copy()
-            header['as_closest_canonical'] = self.as_closest_canonical
-            ndim = img.header['dim'][0]
-            spatial_rank = min(ndim, 3)
-            header['spatial_shape'] = img.header['dim'][1:spatial_rank + 1]
-
-            if self.as_closest_canonical:
-                img = nib.as_closest_canonical(img)
-                header['affine'] = img.affine
-
-            img_array.append(np.array(img.get_fdata(dtype=self.dtype)))
-            img.uncache()
-
-            if self.image_only:
-                continue
-
-            if not compatible_meta:
-                for meta_key in header:
-                    meta_datum = header[meta_key]
-                    if type(meta_datum).__name__ == 'ndarray' \
-                            and np_str_obj_array_pattern.search(meta_datum.dtype.str) is not None:
-                        continue
-                    compatible_meta[meta_key] = meta_datum
-            else:
-                assert np.allclose(header['affine'], compatible_meta['affine']), \
-                    'affine data of all images should be same.'
-
-        img_array = np.stack(img_array, axis=0) if len(img_array) > 1 else img_array[0]
-        if self.image_only:
-            return img_array
-        return img_array, compatible_meta
-
-
-class LoadPNG(Transform):
-    """
-    Load common 2D image format (PNG, JPG, etc. using PIL) file or files from provided path.
-    It's based on the Image module in PIL library.
-    """
-
-    def __init__(self, dtype=np.float32):
-        """Args:
-            dtype (np.dtype, optional): if not None convert the loaded image to this data type.
-
-        """
-        self.dtype = dtype
-
-    def __call__(self, filename):
-        """
-        Args:
-            filename (str, list, tuple, file): path file or file-like object or a list of files.
-        """
-        filename = ensure_tuple(filename)
-        img_array = list()
-        for name in filename:
-            img = np.asarray(Image.open(name))
-            if self.dtype:
-                img = img.astype(self.dtype)
-            img_array.append(img)
-
-        return np.stack(img_array, axis=0) if len(img_array) > 1 else img_array[0]
-
-
-class AsChannelFirst(Transform):
-    """
-    Change the channel dimension of the image to the first dimension.
-
-    Most of the image transformations in ``monai.transforms``
-    assume the input image is in the channel-first format, which has the shape
-    (num_channels, spatial_dim_1[, spatial_dim_2, ...]).
-
-    This transform could be used to convert, for example, a channel-last image array in shape
-    (spatial_dim_1[, spatial_dim_2, ...], num_channels) into the channel-first format,
-    so that the multidimensional image array can be correctly interpreted by the other transforms.
-
-    Args:
-        channel_dim (int): which dimension of input image is the channel, default is the last dimension.
-    """
-
-    def __init__(self, channel_dim=-1):
-        assert isinstance(channel_dim, int) and channel_dim >= -1, 'invalid channel dimension.'
-        self.channel_dim = channel_dim
-
-    def __call__(self, img):
-        return np.moveaxis(img, self.channel_dim, 0)
-
-
-class AsChannelLast(Transform):
-    """
-    Change the channel dimension of the image to the last dimension.
-
-    Some of other 3rd party transforms assume the input image is in the channel-last format with shape
-    (spatial_dim_1[, spatial_dim_2, ...], num_channels).
-
-    This transform could be used to convert, for example, a channel-first image array in shape
-    (num_channels, spatial_dim_1[, spatial_dim_2, ...]) into the channel-last format,
-    so that MONAI transforms can construct a chain with other 3rd party transforms together.
-
-    Args:
-        channel_dim (int): which dimension of input image is the channel, default is the first dimension.
-    """
-
-    def __init__(self, channel_dim=0):
-        assert isinstance(channel_dim, int) and channel_dim >= -1, 'invalid channel dimension.'
-        self.channel_dim = channel_dim
-
-    def __call__(self, img):
-        return np.moveaxis(img, self.channel_dim, -1)
-
-
-class AddChannel(Transform):
-    """
-    Adds a 1-length channel dimension to the input image.
-
-    Most of the image transformations in ``monai.transforms``
-    assumes the input image is in the channel-first format, which has the shape
-    (num_channels, spatial_dim_1[, spatial_dim_2, ...]).
-
-    This transform could be used, for example, to convert a (spatial_dim_1[, spatial_dim_2, ...])
-    spatial image into the channel-first format so that the
-    multidimensional image array can be correctly interpreted by the other
-    transforms.
-    """
-
-    def __call__(self, img):
-        return img[None]
-
-
-class RepeatChannel(Transform):
-    """
-    Repeat channel data to construct expected input shape for models.
-    The `repeats` count includes the origin data, for example:
-    ``RepeatChannel(repeats=2)([[1, 2], [3, 4]])`` generates: ``[[1, 2], [1, 2], [3, 4], [3, 4]]``
-
-    Args:
-        repeats (int): the number of repetitions for each element.
-    """
-
-    def __init__(self, repeats):
-        assert repeats > 0, 'repeats count must be greater than 0.'
-        self.repeats = repeats
-
-    def __call__(self, img):
-        return np.repeat(img, self.repeats, 0)
-
-
-class CastToType(Transform):
-    """
-    Cast the image data to specified numpy data type.
-    """
-
-    def __init__(self, dtype=np.float32):
-        """
-        Args:
-            dtype (np.dtype): convert image to this data type, default is `np.float32`.
-        """
-        self.dtype = dtype
-
-    def __call__(self, img):
-        assert isinstance(img, np.ndarray), 'image must be numpy array.'
-        return img.astype(self.dtype)
-
-
-class ToTensor(Transform):
-    """
-    Converts the input image to a tensor without applying any other transformations.
-    """
-
-    def __call__(self, img):
-        if torch.is_tensor(img):
-            return img.contiguous()
-        return torch.as_tensor(np.ascontiguousarray(img))
-
-
-class Transpose(Transform):
-    """
-    Transposes the input image based on the given `indices` dimension ordering.
-    """
-
-    def __init__(self, indices):
-        self.indices = indices
-
-    def __call__(self, img):
-        return img.transpose(self.indices)
-
-
-class RandGaussianNoise(Randomizable, Transform):
-    """Add Gaussian noise to image.
-
-    Args:
-        prob (float): Probability to add Gaussian noise.
-        mean (float or array of floats): Mean or “centre” of the distribution.
-        std (float): Standard deviation (spread) of distribution.
-    """
-
-    def __init__(self, prob=0.1, mean=0.0, std=0.1):
-        self.prob = prob
-        self.mean = mean
-        self.std = std
-        self._do_transform = False
-        self._noise = None
-
-    def randomize(self, im_shape):
-        self._do_transform = self.R.random() < self.prob
-        self._noise = self.R.normal(self.mean, self.R.uniform(0, self.std), size=im_shape)
-
-    def __call__(self, img):
-        self.randomize(img.shape)
-        return img + self._noise if self._do_transform else img
 
 
 class Flip(Transform):
@@ -621,251 +366,6 @@ class Zoom(Transform):
         return zoomed[tuple(slice_vec)]
 
 
-class ShiftIntensity(Transform):
-    """Shift intensity uniformly for the entire image with specified `offset`.
-
-    Args:
-        offset (int or float): offset value to shift the intensity of image.
-    """
-
-    def __init__(self, offset):
-        self.offset = offset
-
-    def __call__(self, img):
-        return img + self.offset
-
-
-class RandShiftIntensity(Randomizable, Transform):
-    """Randomly shift intensity with randomly picked offset.
-    """
-
-    def __init__(self, offsets, prob=0.1):
-        """
-        Args:
-            offsets(int, float, tuple or list): offset range to randomly shift.
-                if single number, offset value is picked from (-offsets, offsets).
-            prob (float): probability of shift.
-        """
-        self.offsets = (-offsets, offsets) if not isinstance(offsets, (list, tuple)) else offsets
-        assert len(self.offsets) == 2, 'offsets should be a number or pair of numbers.'
-        self.prob = prob
-        self._do_transform = False
-
-    def randomize(self):
-        self._offset = self.R.uniform(low=self.offsets[0], high=self.offsets[1])
-        self._do_transform = self.R.random() < self.prob
-
-    def __call__(self, img):
-        self.randomize()
-        if not self._do_transform:
-            return img
-        shifter = ShiftIntensity(self._offset)
-        return shifter(img)
-
-
-class ScaleIntensity(Transform):
-    """
-    Scale the intensity of input image to the given value range (minv, maxv).
-    If `minv` and `maxv` not provided, use `factor` to scale image by ``v = v * (1 + factor)``.
-    """
-
-    def __init__(self, minv=0.0, maxv=1.0, factor=None, dtype=np.float32):
-        """
-        Args:
-            minv (int or float): minimum value of output data.
-            maxv (int or float): maximum value of output data.
-            factor (float): factor scale by ``v = v * (1 + factor)``.
-            dtype (np.dtype): expected output data type.
-        """
-        self.minv = minv
-        self.maxv = maxv
-        self.factor = factor
-        self.dtype = dtype
-
-    def __call__(self, img):
-        if self.minv is not None and self.maxv is not None:
-            return rescale_array(img, self.minv, self.maxv, self.dtype)
-        else:
-            return (img * (1 + self.factor)).astype(self.dtype)
-
-
-class RandScaleIntensity(Randomizable, Transform):
-    """
-    Randomly scale the intensity of input image by ``v = v * (1 + factor)`` where the `factor`
-    is randomly picked from (factors[0], factors[0]).
-    """
-
-    def __init__(self, factors, prob=0.1, dtype=np.float32):
-        """
-        Args:
-            factors(float, tuple or list): factor range to randomly scale by ``v = v * (1 + factor)``.
-                if single number, factor value is picked from (-factors, factors).
-            prob (float): probability of scale.
-            dtype (np.dtype): expected output data type.
-        """
-        self.factors = (-factors, factors) if not isinstance(factors, (list, tuple)) else factors
-        assert len(self.factors) == 2, 'factors should be a number or pair of numbers.'
-        self.prob = prob
-        self.dtype = dtype
-        self._do_transform = False
-
-    def randomize(self):
-        self.factor = self.R.uniform(low=self.factors[0], high=self.factors[1])
-        self._do_transform = self.R.random() < self.prob
-
-    def __call__(self, img):
-        self.randomize()
-        if not self._do_transform:
-            return img
-        scaler = ScaleIntensity(minv=None, maxv=None, factor=self.factor, dtype=self.dtype)
-        return scaler(img)
-
-
-class NormalizeIntensity(Transform):
-    """Normalize input based on provided args, using calculated mean and std if not provided
-    (shape of subtrahend and divisor must match. if 0, entire volume uses same subtrahend and
-    divisor, otherwise the shape can have dimension 1 for channels).
-    This transform can normalize only non-zero values or entire image, and can also calculate
-    mean and std on each channel separately.
-
-    Args:
-        subtrahend (ndarray): the amount to subtract by (usually the mean).
-        divisor (ndarray): the amount to divide by (usually the standard deviation).
-        nonzero (bool): whether only normalize non-zero values.
-        channel_wise (bool): if using calculated mean and std, calculate on each channel separately
-            or calculate on the entire image directly.
-    """
-
-    def __init__(self, subtrahend=None, divisor=None, nonzero=False, channel_wise=False):
-        if subtrahend is not None or divisor is not None:
-            assert isinstance(subtrahend, np.ndarray) and isinstance(divisor, np.ndarray), \
-                'subtrahend and divisor must be set in pair and in numpy array.'
-        self.subtrahend = subtrahend
-        self.divisor = divisor
-        self.nonzero = nonzero
-        self.channel_wise = channel_wise
-
-    def _normalize(self, img):
-        slices = (img != 0) if self.nonzero else np.ones(img.shape, dtype=np.bool_)
-        if np.any(slices):
-            if self.subtrahend is not None and self.divisor is not None:
-                img[slices] = (img[slices] - self.subtrahend[slices]) / self.divisor[slices]
-            else:
-                img[slices] = (img[slices] - np.mean(img[slices])) / np.std(img[slices])
-        return img
-
-    def __call__(self, img):
-        if self.channel_wise:
-            for i, d in enumerate(img):
-                img[i] = self._normalize(d)
-        else:
-            img = self._normalize(img)
-
-        return img
-
-
-class ThresholdIntensity(Transform):
-    """Filter the intensity values of whole image to below threshold or above threshold.
-    And fill the remaining parts of the image to the `cval` value.
-
-    Args:
-        threshold (float or int): the threshold to filter intensity values.
-        above (bool): filter values above the threshold or below the threshold, default is True.
-        cval (float or int): value to fill the remaining parts of the image, default is 0.
-    """
-
-    def __init__(self, threshold, above=True, cval=0):
-        assert isinstance(threshold, (float, int)), 'must set the threshold to filter intensity.'
-        self.threshold = threshold
-        self.above = above
-        self.cval = cval
-
-    def __call__(self, img):
-        return np.where(img > self.threshold if self.above else img < self.threshold, img, self.cval)
-
-
-class ScaleIntensityRange(Transform):
-    """Apply specific intensity scaling to the whole numpy array.
-    Scaling from [a_min, a_max] to [b_min, b_max] with clip option.
-
-    Args:
-        a_min (int or float): intensity original range min.
-        a_max (int or float): intensity original range max.
-        b_min (int or float): intensity target range min.
-        b_max (int or float): intensity target range max.
-        clip (bool): whether to perform clip after scaling.
-    """
-
-    def __init__(self, a_min, a_max, b_min, b_max, clip=False):
-        self.a_min = a_min
-        self.a_max = a_max
-        self.b_min = b_min
-        self.b_max = b_max
-        self.clip = clip
-
-    def __call__(self, img):
-        img = (img - self.a_min) / (self.a_max - self.a_min)
-        img = img * (self.b_max - self.b_min) + self.b_min
-        if self.clip:
-            img = np.clip(img, self.b_min, self.b_max)
-
-        return img
-
-
-class AdjustContrast(Transform):
-    """Changes image intensity by gamma. Each pixel/voxel intensity is updated as:
-        `x = ((x - min) / intensity_range) ^ gamma * intensity_range + min`
-
-    Args:
-        gamma (float): gamma value to adjust the contrast as function.
-    """
-
-    def __init__(self, gamma):
-        assert isinstance(gamma, (float, int)), 'gamma must be a float or int number.'
-        self.gamma = gamma
-
-    def __call__(self, img):
-        epsilon = 1e-7
-        img_min = img.min()
-        img_range = img.max() - img_min
-        return np.power(((img - img_min) / float(img_range + epsilon)), self.gamma) * img_range + img_min
-
-
-class RandAdjustContrast(Randomizable, Transform):
-    """Randomly changes image intensity by gamma. Each pixel/voxel intensity is updated as:
-        `x = ((x - min) / intensity_range) ^ gamma * intensity_range + min`
-
-    Args:
-        prob (float): Probability of adjustment.
-        gamma (tuple of float or float): Range of gamma values.
-            If single number, value is picked from (0.5, gamma), default is (0.5, 4.5).
-    """
-
-    def __init__(self, prob=0.1, gamma=(0.5, 4.5)):
-        self.prob = prob
-        if not isinstance(gamma, (tuple, list)):
-            assert gamma > 0.5, \
-                'if gamma is single number, must greater than 0.5 and value is picked from (0.5, gamma)'
-            self.gamma = (0.5, gamma)
-        else:
-            self.gamma = gamma
-        assert len(self.gamma) == 2, 'gamma should be a number or pair of numbers.'
-
-        self._do_transform = False
-        self.gamma_value = None
-
-    def randomize(self):
-        self._do_transform = self.R.random_sample() < self.prob
-        self.gamma_value = self.R.uniform(low=self.gamma[0], high=self.gamma[1])
-
-    def __call__(self, img):
-        self.randomize()
-        if not self._do_transform:
-            return img
-        adjuster = AdjustContrast(self.gamma_value)
-        return adjuster(img)
-
-
 class Rotate90(Transform):
     """
     Rotate an array by 90 degrees in the plane specified by `axes`.
@@ -927,180 +427,6 @@ class RandRotate90(Randomizable, Transform):
             return img
         rotator = Rotate90(self._rand_k, self.spatial_axes)
         return rotator(img)
-
-
-class SpatialPad(Transform):
-    """Performs padding to the data, symmetric for all sides or all on one side for each dimension.
-     Uses np.pad so in practice, a mode needs to be provided. See numpy.lib.arraypad.pad
-     for additional details.
-
-    Args:
-        spatial_size (list): the spatial size of output data after padding.
-        method (str): pad image symmetric on every side or only pad at the end sides. default is 'symmetric'.
-        mode (str): one of the following string values or a user supplied function: {'constant', 'edge', 'linear_ramp',
-            'maximum', 'mean', 'median', 'minimum', 'reflect', 'symmetric', 'wrap', 'empty', <function>}
-            for more details, please check: https://docs.scipy.org/doc/numpy/reference/generated/numpy.pad.html
-    """
-
-    def __init__(self, spatial_size, method='symmetric', mode='constant'):
-        assert isinstance(spatial_size, (list, tuple)), 'spatial_out_size must be list or tuple.'
-        self.spatial_size = spatial_size
-        assert method in ('symmetric', 'end'), 'unsupported padding type.'
-        self.method = method
-        assert isinstance(mode, str), 'mode must be str.'
-        self.mode = mode
-
-    def _determine_data_pad_width(self, data_shape):
-        if self.method == 'symmetric':
-            pad_width = list()
-            for i in range(len(self.spatial_size)):
-                width = max(self.spatial_size[i] - data_shape[i], 0)
-                pad_width.append((width // 2, width - (width // 2)))
-            return pad_width
-        else:
-            return [(0, max(self.spatial_size[i] - data_shape[i], 0)) for i in range(len(self.spatial_size))]
-
-    def __call__(self, img):
-        data_pad_width = self._determine_data_pad_width(img.shape[1:])
-        all_pad_width = [(0, 0)] + data_pad_width
-        img = np.pad(img, all_pad_width, self.mode)
-        return img
-
-
-class SpatialCrop(Transform):
-    """General purpose cropper to produce sub-volume region of interest (ROI).
-    It can support to crop ND spatial (channel-first) data.
-    Either a spatial center and size must be provided, or alternatively if center and size
-    are not provided, the start and end coordinates of the ROI must be provided.
-    The sub-volume must sit the within original image.
-    Note: This transform will not work if the crop region is larger than the image itself.
-    """
-
-    def __init__(self, roi_center=None, roi_size=None, roi_start=None, roi_end=None):
-        """
-        Args:
-            roi_center (list or tuple): voxel coordinates for center of the crop ROI.
-            roi_size (list or tuple): size of the crop ROI.
-            roi_start (list or tuple): voxel coordinates for start of the crop ROI.
-            roi_end (list or tuple): voxel coordinates for end of the crop ROI.
-        """
-        if roi_center is not None and roi_size is not None:
-            roi_center = np.asarray(roi_center, dtype=np.uint16)
-            roi_size = np.asarray(roi_size, dtype=np.uint16)
-            self.roi_start = np.subtract(roi_center, np.floor_divide(roi_size, 2))
-            self.roi_end = np.add(self.roi_start, roi_size)
-        else:
-            assert roi_start is not None and roi_end is not None, 'roi_start and roi_end must be provided.'
-            self.roi_start = np.asarray(roi_start, dtype=np.uint16)
-            self.roi_end = np.asarray(roi_end, dtype=np.uint16)
-
-        assert np.all(self.roi_start >= 0), 'all elements of roi_start must be greater than or equal to 0.'
-        assert np.all(self.roi_end > 0), 'all elements of roi_end must be positive.'
-        assert np.all(self.roi_end >= self.roi_start), 'invalid roi range.'
-
-    def __call__(self, img):
-        max_end = img.shape[1:]
-        sd = min(len(self.roi_start), len(max_end))
-        assert np.all(max_end[:sd] >= self.roi_start[:sd]), 'roi start out of image space.'
-        assert np.all(max_end[:sd] >= self.roi_end[:sd]), 'roi end out of image space.'
-
-        slices = [slice(None)] + [slice(s, e) for s, e in zip(self.roi_start[:sd], self.roi_end[:sd])]
-        return img[tuple(slices)]
-
-
-class CenterSpatialCrop(Transform):
-    """
-    Crop at the center of image with specified ROI size.
-
-    Args:
-        roi_size (list, tuple): the spatial size of the crop region e.g. [224,224,128]
-    """
-
-    def __init__(self, roi_size):
-        self.roi_size = roi_size
-
-    def __call__(self, img):
-        center = [i // 2 for i in img.shape[1:]]
-        cropper = SpatialCrop(roi_center=center, roi_size=self.roi_size)
-        return cropper(img)
-
-
-class RandSpatialCrop(Randomizable, Transform):
-    """
-    Crop image with random size or specific size ROI. It can crop at a random position as center
-    or at the image center. And allows to set the minimum size to limit the randomly generated ROI.
-    This transform assumes all the expected fields specified by `keys` have same shape.
-
-    Args:
-        roi_size (list, tuple): if `random_size` is True, the spatial size of the minimum crop region.
-            if `random_size` is False, specify the expected ROI size to crop. e.g. [224, 224, 128]
-        random_center (bool): crop at random position as center or the image center.
-        random_size (bool): crop with random size or specific size ROI.
-    """
-
-    def __init__(self, roi_size, random_center=True, random_size=True):
-        self.roi_size = roi_size
-        self.random_center = random_center
-        self.random_size = random_size
-
-    def randomize(self, img_size):
-        self._size = [self.roi_size] * len(img_size) if not isinstance(self.roi_size, (list, tuple)) else self.roi_size
-        if self.random_size:
-            self._size = [self.R.randint(low=self._size[i], high=img_size[i] + 1) for i in range(len(img_size))]
-        if self.random_center:
-            valid_size = get_valid_patch_size(img_size, self._size)
-            self._slices = ensure_tuple(slice(None)) + get_random_patch(img_size, valid_size, self.R)
-
-    def __call__(self, img):
-        self.randomize(img.shape[1:])
-        if self.random_center:
-            return img[self._slices]
-        else:
-            cropper = CenterSpatialCrop(self._size)
-            return cropper(img)
-
-
-class CropForeground(Transform):
-    """
-    Crop an image using a bounding box. The bounding box is generated by selecting foreground using select_fn
-    at channels channel_indexes. margin is added in each spatial dimension of the bounding box.
-    The typical usage is to help training and evaluation if the valid part is small in the whole medical image.
-    Users can define arbitrary function to select expected foreground from the whole image or specified channels.
-    And it can also add margin to every dim of the bounding box of foreground object.
-    For example:
-
-    .. code-block:: python
-
-        image = np.array(
-            [[[0, 0, 0, 0, 0],
-              [0, 1, 2, 1, 0],
-              [0, 1, 3, 2, 0],
-              [0, 1, 2, 1, 0],
-              [0, 0, 0, 0, 0]]])  # 1x5x5, single channel 5x5 image
-        cropper = CropForeground(select_fn=lambda x: x > 1, margin=0)
-        print(cropper(image))
-        [[[2, 1],
-          [3, 2],
-          [2, 1]]]
-
-    """
-
-    def __init__(self, select_fn=lambda x: x > 0, channel_indexes=None, margin=0):
-        """
-        Args:
-            select_fn (Callable): function to select expected foreground, default is to select values > 0.
-            channel_indexes (int, tuple or list): if defined, select foregound only on the specified channels
-                of image. if None, select foreground on the whole image.
-            margin (int): add margin to all dims of the bounding box.
-        """
-        self.select_fn = select_fn
-        self.channel_indexes = ensure_tuple(channel_indexes) if channel_indexes is not None else None
-        self.margin = margin
-
-    def __call__(self, img):
-        box_start, box_end = generate_spatial_bounding_box(img, self.select_fn, self.channel_indexes, self.margin)
-        cropper = SpatialCrop(roi_start=box_start, roi_end=box_end)
-        return cropper(img)
 
 
 class RandRotate(Randomizable, Transform):
@@ -1738,26 +1064,3 @@ class Rand3DElastic(Randomizable, Transform):
             grid[:3] += gaussian(self.rand_offset[None])[0] * self.magnitude
             grid = self.rand_affine_grid(grid=grid)
         return self.resampler(img, grid, mode)
-
-
-class SqueezeDim(Transform):
-    """
-    Squeeze undesired unitary dimensions
-    """
-
-    def __init__(self, dim=None):
-        """
-        Args:
-            dim (int): dimension to be squeezed.
-                Default: None (all dimensions of size 1 will be removed)
-        """
-        if dim is not None:
-            assert isinstance(dim, int) and dim >= -1, 'invalid channel dimension.'
-        self.dim = dim
-
-    def __call__(self, img):
-        """
-        Args:
-            img (ndarray): numpy arrays with required dimension `dim` removed
-        """
-        return np.squeeze(img, self.dim)
